feat: improve setting menu a11y
